3.15. добавляет `get-mongo-connection-string.ts`
