+++ project/libs/shared/helpers/src/index.ts
@@ -1,1 +1,1 @@
-// export * from './lib/helpers';
+export * from './lib/get-mongo-connection-string';
